Update apt-get install
Per best practices outlined here:
https://docs.docker.com/articles/dockerfile_best-practices/

In particular:
> Aside from the `ADD` and `COPY` commands cache checking will not look
at the files in the container to determine a cache match. For example,
when processing a `RUN apt-get -y update` command the files updated in
the container will not be examined to determine if a cache hit exists.
In that case just the command string itself will be used to find a
match.